article-downloader-image/content/requirements.txt
redis==2.10.6
newspaper3k==0.2.8